backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
scikit-learn
xgboost
pandas